pydantic/requirements.txt
pydantic-ai
gradio
python-dotenv
openai
